requirements.txt
numpy
scipy
matplotlib
openpyxl
pandas
scikit-learn
pygame
pathfinding
neat-python